refactor: reorganize component imports to layout directory
- Updated import paths for various components, moving them from their previous locations to a new `layout` directory for better structure and consistency.
- This includes changes to imports in filters, legal, resources, landing page, and layout components.
